-more fixes to sec sys %)
git-svn-id: svn://300murlocs.com/core/stable/code@9 b8eaa9ba-51a4-de11-8768-001a646d8106

+++ src/game/Chat.cpp
@@ -666,7 +666,7 @@ ChatCommand * ChatHandler::getCommandTable()
                 Field *fields = result->Fetch();
                 std::string name = fields[0].GetCppString();
 
-                SetDataForCommandInTable(commandTable, name.c_str(), fields[1].GetUInt16(), fields[2].GetCppString(), name);
+                SetDataForCommandInTable(commandTable, name.c_str(), fields[1].GetUInt32(), fields[2].GetCppString(), name);
 
             } while(result->NextRow());
             delete result;
